Added logic for reading url
"pathName" stores the result from the URL and then check that value
against "/" or "/overview" in the first "if" block. if the result is
true then a response is sent to the client otherwise checked is moved to
the next "else if" statement. Finally if both "if" checks fails then a
"404" response is sent.