Generating html template using JSON
Html template is generated running a ".map()" method which calls
"replaceTemplate" function. The function takes 2 arguments "tempCard" and "ele", these arguments are passed inside the function and then the function replaces place holders with the current value of "ele". Inside the function there is "output" variables which appends the old "output" to the new "output".Finally the function returns "output" and a "join()" is run to combine all the values returned.